#6 Implementación Inicial
Hice un Refactoreo del funcionamiento de las combinaciones, agregé el item [Piedra con baba] dentro de la lista y finalmente implemente la interacción.

+++ Assets/Scripts/Core.Interaction/Ignition/Slime.cs
@@ -194,7 +194,7 @@ public class Slime : Item, IIgnitableObject
 
         if (inventory != null)
         {
-            if (inventory.equiped != null && inventory.equiped.ID == (int)ItemID.Antorcha)
+            if (inventory.equiped != null && inventory.equiped.ID == ItemID.Antorcha)
             {
                 Torch torch = (Torch)inventory.equiped;
                 if (torch.isBurning)
